+++ test/HFunds.test.js
@@ -85,9 +85,9 @@ contract('Funds', function([_, user, someone]) {
       balanceProxy = await tracker(this.proxy.address);
     });
 
-    it('normal', async function() {
-      const token = [this.token0.address, this.token1.address];
-      const value = [ether('100'), ether('200')];
+    it('inject', async function() {
+      const token = [this.token0.address];
+      const value = [ether('100')];
       const to = this.hFunds.address;
       const data = abi.simpleEncode('updateTokens(address[])', token);
       await this.token0.transfer(this.proxy.address, value[0], {
@@ -124,10 +124,43 @@ contract('Funds', function([_, user, someone]) {
       profileGas(receipt);
     });
 
-    it('native token - zero address', async function() {
-      const token = [this.token0.address, ZERO_ADDRESS];
-      const msgValue = ether('0.1');
-      const value = [ether('200'), ether('1')];
+
+
+    it('add funds', async function() {
+      const token = [this.token0.address];
+      const value = [ether('100')];
+      const to = this.hFunds.address;
+      const data = abi.simpleEncode(
+        'addFunds(address[],uint256[])',
+        token,
+        value
+      );
+      await this.token0.transfer(user, value[0], {
+        from: providerAddresses[0],
+      });
+      await this.token0.approve(this.proxy.address, value[0], { from: user });
+
+      const receipt = await this.proxy.execMock(to, data, {
+        from: user,
+        value: ether('0.1'),
+      });
+
+      await expectEvent.inTransaction(receipt.tx, this.token0, 'Transfer', {
+        from: user,
+        to: this.proxy.address,
+        value: value[0],
+      });
+      await expectEvent.inTransaction(receipt.tx, this.token0, 'Transfer', {
+        from: this.proxy.address,
+        to: user,
+        value: value[0],
+      });
+      profileGas(receipt);
+    });
+
+    it('USDT', async function() {
+      const token = [this.usdt.address];
+      const value = [new BN('1000000')];
       const to = this.hFunds.address;
       const data = abi.simpleEncode('updateTokens(address[])', token);
       // Transfer tokens to proxy first
@@ -162,10 +195,61 @@ contract('Funds', function([_, user, someone]) {
       profileGas(receipt);
     });
 
-    it('native token - 0xEEEE', async function() {
-      const token = [this.token0.address, NATIVE_TOKEN_ADDRESS];
-      const msgValue = ether('0.1');
-      const value = [ether('200'), ether('1')];
+  describe('multiple tokens', function() {
+    before(async function() {
+      this.token0 = await IToken.at(tokenAddresses[0]);
+      this.token1 = await IToken.at(tokenAddresses[1]);
+    });
+    it('inject', async function() {
+      const token = [this.token0.address, this.token1.address];
+      const value = [ether('100'), ether('100')];
+      const to = this.hFunds.address;
+      const data = abi.simpleEncode(
+        'inject(address[],uint256[])',
+        token,
+        value
+      );
+      await this.token0.transfer(user, value[0], {
+        from: providerAddresses[0],
+      });
+      await this.token0.approve(this.proxy.address, value[0], { from: user });
+      await this.token1.transfer(user, value[1], {
+        from: providerAddresses[1],
+      });
+      await this.token1.approve(this.proxy.address, value[1], { from: user });
+
+      const receipt = await this.proxy.execMock(to, data, {
+        from: user,
+        value: ether('1'),
+      });
+
+      await expectEvent.inTransaction(receipt.tx, this.token0, 'Transfer', {
+        from: user,
+        to: this.proxy.address,
+        value: value[0],
+      });
+      await expectEvent.inTransaction(receipt.tx, this.token0, 'Transfer', {
+        from: this.proxy.address,
+        to: user,
+        value: value[0],
+      });
+
+      await expectEvent.inTransaction(receipt.tx, this.token1, 'Transfer', {
+        from: user,
+        to: this.proxy.address,
+        value: value[1],
+      });
+      await expectEvent.inTransaction(receipt.tx, this.token1, 'Transfer', {
+        from: this.proxy.address,
+        to: user,
+        value: value[1],
+      });
+      profileGas(receipt);
+    });
+
+    it('add funds', async function() {
+      const token = [this.token0.address, this.token1.address];
+      const value = [ether('100'), ether('100')];
       const to = this.hFunds.address;
       const data = abi.simpleEncode('updateTokens(address[])', token);
       // Transfer tokens to proxy first